fix: rename Environment from context
